% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{ewma}
\alias{ewma}
\title{Exponentially weighted moving average}
\usage{
ewma(x, lambda = 0.94)
}
\arguments{
\item{x}{a numeric vector of asset returns.}

\item{lambda}{decay parameter for calculation of weights; default is 0.94.}
}
\description{
Estimates volatilities of a return series by means of an exponentially weighted moving average.
}
