% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsim.R
\name{hsim}
\alias{hsim}
\title{Nonparametric calculation of Value at Risk and Expected Shortfall}
\usage{
hsim(x, p = 0.95, method = c("age", "basic"), lambda = 0.98)
}
\arguments{
\item{x}{a numeric vector of asset returns.}

\item{p}{confidence level for VaR calculation; default is 0.95%.}

\item{method}{method to be used for calculation; default is "basic".}

\item{lambda}{decay parameter for calculation of weights; default is 0.98.}
}
\description{
Calculates Value at Risk and Expected Shortfall by means of plain and age weighted historical simulation
}
