% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.quarks.R
\name{plot.quarks}
\alias{plot.quarks}
\title{Plot Method for the Package 'quarks'}
\usage{
\method{plot}{quarks}(x, ...)
}
\arguments{
\item{x}{an input object of class \emph{quarks}.}

\item{...}{additional arguments of the standard plot method.}
}
\value{
None
}
\description{
This function regulates how objects created by the package quarks are
plotted.
}
