% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quarks.R
\name{print.quarks}
\alias{print.quarks}
\title{Print Method for the Package 'quarks'}
\usage{
\method{print}{quarks}(x, ...)
}
\arguments{
\item{x}{an input object of class \code{quarks}.}

\item{...}{included for compatibility; additional arguments will however
not affect the output.}
}
\value{
None
}
\description{
This function regulates how objects created by the package \code{quarks} are
printed.
}
