% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vwhsim.R
\name{vwhsim}
\alias{vwhsim}
\title{Volatility weighted historical simulation}
\usage{
vwhsim(x, p = 0.95, lambda = 0.94)
}
\arguments{
\item{x}{a numeric vector of asset returns.}

\item{p}{confidence level for VaR calculation; default is 0.95%.}

\item{lambda}{decay parameter for calculation of weights; default is 0.94.}
}
\description{
Calculates Value at Risk and Expected Shortfall by means of volatility weighted historical simulation. Volatilities
are estimated with an exponentially weighed moving average.
}
