% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs.R
\name{fhs}
\alias{fhs}
\title{Filtered historical simulation}
\usage{
fhs(x, p = 0.975, lambda = 0.94, nboot = 10000, nahead = 1, arma = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector of asset returns}

\item{p}{confidence level for VaR calculation; default is 0.975}

\item{lambda}{decay factor for the calculation of weights; default is 0.94}

\item{nboot}{size of bootstrap sample; default is 10000}

\item{nahead}{n-day ahead forecasts of VaR and ES; default is 1}

\item{arma}{fitting an ARMA-model to the return series before volatility
filtering; default is FALSE}

\item{...}{additional arguments of the \emph{arima} function}
}
\value{
Returns a list with the following elements:
\describe{
\item{VaR}{Calculated Value at Risk}
\item{ES}{Calculated Expected Shortfall (Conditional Value at Risk)}
\item{armacoefs}{Estimated model parameters}
}
}
\description{
Calculates univariate Value at Risk and Expected Shortfall (also called
Conditional Value at Risk) by means of filtered historical simulation.
Volatility is estimated with an exponentially weighted moving
average.
}
\examples{
prices <- DAX30$price.close
returns <- diff(log(prices))
fhs(x = returns, p = 0.975, lambda = 0.94, nahead = 5)
}
